refactor: minor replace from navbar
