+++ src/components/authorization/authorization.scss
@@ -1,5 +1,14 @@
 @import "./../../css/vars.scss";
 
+.auth-panel-container {
+  display: flex;
+  gap: 10px;
+  .navbar-name {
+    min-width: 60px;
+    text-align: end;
+  }
+}
+
 .auth-form {
   max-width: 500px;
   padding: 50px;
@@ -23,7 +32,6 @@
     width: 100%;
   }
 }
-
 .auth__warn {
   color: red;
   visibility: hidden;
