+++ src/components/authorization/sign-out.ts
@@ -2,7 +2,7 @@ import { openAuthModal } from './authorization';
 
 function signOut() {
   const headerButton = document.querySelector('.navbar-auth') as HTMLFormElement;
-  const headerName = document.querySelector('.navbar .navbar-name') as HTMLSpanElement;
+  const headerName = document.querySelector('.navbar-name') as HTMLSpanElement;
 
   localStorage.clear();
   headerButton.innerText = 'Войти';
