+++ src/components/header/header.scss
@@ -1,3 +1,6 @@
 header {
   background-color: blanchedalmond;
+  display: flex;
+  align-items: center;
+  justify-content: space-evenly;
 }
